Refactored Precision and Recall functions into _BasePrecisionRecallSupport (child of _BaseClassification). Refactored Accuracy into _BaseClassification.
